[Feat] 서비스 그룹 목록 조회 api 연동
